Grudgingly add functions to manage gameobjects in the network manager
This is straight up wrong in my very unprofessional opinion. I don't like the network manager keeping track of all the game objects. I don't have a better solution at the moment and I want to keep it close to the book for now so I will leave it as is...